requirements.txt
numpy==1.26.4
streamlit==1.40.0
pillow==10.3.0